requirements-monitoring.txt
# Core Monitoring Dependencies
psutil>=5.9.0        # System metrics collection
redis>=4.5.0         # Cache and data storage
aiohttp>=3.8.0       # Async HTTP client
websockets>=10.4     # WebSocket support
prometheus_client>=0.16.0  # Metrics export

# Database
psycopg2-binary>=2.9.5  # PostgreSQL adapter
SQLAlchemy>=1.4.0    # SQL toolkit and ORM
alembic>=1.9.0       # Database migrations

# Data Processing
numpy>=1.23.0        # Numerical computations
pandas>=1.5.0        # Data analysis
scikit-learn>=1.2.0  # Machine learning for anomaly detection

# Visualization
plotly>=5.13.0       # Interactive plots
dash>=2.8.0          # Dashboard framework
dash-bootstrap-components>=1.4.0  # Bootstrap components

# Alerting
slackclient>=2.9.4   # Slack integration
aiosmtplib>=2.0.0    # Async SMTP client
jinja2>=3.1.0        # Template engine
markdown>=3.4.0      # Markdown support

# Caching and Compression
gzip-stream>=1.1.0   # Streaming gzip compression
lz4>=4.3.0          # Fast compression
msgpack>=1.0.4       # Efficient serialization

# Security
cryptography>=39.0.0  # Cryptographic recipes
PyJWT>=2.6.0         # JSON Web Tokens
bcrypt>=4.0.0        # Password hashing

# Time Series
influxdb-client>=1.36.0  # Time series database
croniter>=1.3.0      # Cron expression parser

# Logging
structlog>=23.1.0    # Structured logging
python-json-logger>=2.0.0  # JSON log formatter

# Testing and Development
pytest>=7.3.0        # Testing framework
pytest-asyncio>=0.21.0  # Async test support
pytest-cov>=4.0.0    # Coverage reporting
black>=23.1.0        # Code formatting
flake8>=6.0.0        # Code linting
mypy>=1.0.0          # Static type checking

# Documentation
Sphinx>=6.1.0        # Documentation generator
sphinx-rtd-theme>=1.2.0  # Documentation theme
myst-parser>=1.0.0   # Markdown support

# Performance Monitoring
py-spy>=0.3.0        # Sampling profiler
memory-profiler>=0.61.0  # Memory profiling
line-profiler>=4.0.0  # Line-by-line profiling

# System Integration
systemd-python>=235   # Systemd integration
supervisor>=4.2.0     # Process control
docker>=6.0.0         # Docker integration

# Networking
aiohttp>=3.8.0       # Async HTTP
httpx>=0.24.0        # HTTP client
dnspython>=2.3.0     # DNS toolkit

# Metrics Collection
statsd>=4.0.0        # StatsD client
dogstatsd>=0.10.0    # DataDog StatsD
pyformance>=0.4      # Metrics library

# Tracing
opentelemetry-api>=1.17.0      # OpenTelemetry API
opentelemetry-sdk>=1.17.0      # OpenTelemetry SDK
opentelemetry-instrumentation>=0.39b0  # Auto-instrumentation

# Health Checks
healthcheck>=1.3.0   # Health check library
watchdog>=3.0.0      # Filesystem monitoring

# Configuration
python-dotenv>=1.0.0  # Environment variables
PyYAML>=6.0          # YAML support
configargparse>=1.5.3 # Configuration parsing

# Backup and Recovery
boto3>=1.26.0        # AWS SDK
google-cloud-storage>=2.7.0  # Google Cloud Storage
azure-storage-blob>=12.14.0  # Azure Blob Storage

# Async Support
asyncio>=3.4.3       # Async I/O
aiofiles>=23.1.0     # Async file operations
aiodns>=3.0.0        # Async DNS resolver

# CLI Tools
click>=8.1.3         # Command line interface
rich>=13.3.0         # Rich text and formatting
tqdm>=4.65.0         # Progress bars

# Version Constraints
setuptools>=65.5.1
wheel>=0.40.0
pip>=23.0.1
